requirements.txt
pillow == 9.5.0
numpy == 1.25.0
argparse == 1.1
PyQt6 == 6.7.0
rembg == 2.0.56
